Scale touch events properly.
Javascript produces touches events with a value of [0,1], whereas the
emulator is expecting linux input codes for scale.

We now properly scale the events and some tests to make sure we scale
properly.

We release this as a beta as it includes a logging message for touch
events that can be used to validate behavior on iPhone/iPad.