Fix touch event serialization issues.
It wass possible for touch events to calculate floats instead of ints,
resulting in serialization failures when making gRPC calls.

We now introduce a set of additional tests that validate that
serialization works before making a remote call, as well as a series of
tests to validate touch scaling.

The unit tests expose the serialization bug.